Fourth Commit | Ms al mismo nivel que ms-ventas *el script de creacion de tablas puede fallar por la relación que requiere al ms-ventas (probar)*

+++ src/main/java/cl/kiosko/ms_pagos/DTO/MetodoPagoResponseDTO.java
@@ -8,6 +8,6 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 public class MetodoPagoResponseDTO {
-    private Long metodoPagoId;
+    private Integer metodoPagoId;
     private String tipoPago;
 }
